refactor: v2를 위해 인터페이스 구조를 맞추기 위한 리팩토링
