+++ README.md
@@ -19,3 +19,7 @@
 - [x] 실행 결과 입력받고 출력하기
 - [x] 사람 별로 사디리 게임 실행 후 결과 구하기
 - [x] 결과를 보고 싶은 사람 받아 실행 결과 보여주기
+
+## Step4 사다리(리팩토링)
+- [ ] Line 을 다른 방식으로 구현하기 (LadderLine)
+- [ ] LadderLine 기존 인터페이스에 조합해서 끼워넣기
