+++ src/main/java/ladder/domain/Ladder.java
@@ -1,5 +1,8 @@
 package ladder.domain;
 
+import ladder.domain.factory.ILine;
+import ladder.domain.factory.LineFactory;
+
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.IntStream;
@@ -7,7 +10,7 @@ import java.util.stream.IntStream;
 import static java.util.stream.Collectors.toList;
 
 public class Ladder {
-    private final List<Line> lines;
+    private final List<ILine> lines;
 
     public Ladder(int playerCount, int height, LineFactory lineFactory) {
         this.lines = IntStream.range(0, height)
@@ -15,11 +18,11 @@ public class Ladder {
                 .collect(toList());
     }
 
-    Ladder(List<Line> lines) {
+    Ladder(List<ILine> lines) {
         this.lines = lines;
     }
 
-    public List<Line> lines() {
+    public List<ILine> lines() {
         return lines;
     }
 
@@ -40,7 +43,7 @@ public class Ladder {
 
     public Players goDown(Players players) {
         Players result = players;
-        for (Line line: lines) {
+        for (ILine line: lines) {
             result = result.move(line);
         }
         return result;
