--- src/main/java/ladder/domain/LineFactory.java
@@ -1,7 +1,0 @@
-package ladder.domain;
-
-@FunctionalInterface
-public interface LineFactory {
-
-    Line create(int playerCount);
-}
