+++ src/main/java/ladder/domain/factory/LineFactory.java
@@ -1,0 +1,7 @@
+package ladder.domain.factory;
+
+@FunctionalInterface
+public interface LineFactory {
+
+    ILine create(int playerCount);
+}
