+++ src/test/java/ladder/domain/LineTest.java
@@ -1,5 +1,6 @@
 package ladder.domain;
 
+import ladder.domain.v1.Line;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
